feat: [PAYMCLOUD-35] Add Opsgenie integration and update module versions (#2535)
* Add Opsgenie integration and update module versions

Integrated Opsgenie as a new monitoring action group and linked it to the AKS alerts. Also updated several module versions to maintain compatibility and ensure newer features and fixes are included in the project.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Fix markdown formatting in README.md with pre-commit

Replaced HTML line breaks with cleaner Markdown formatting. This improves the readability and consistency of the documentation.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Format DNS configuration entries uniformly

Aligned the equal signs in the DNS configuration section to improve readability and maintain consistency in the `terraform.tfvars` file. This change does not affect any values but enhances the clarity of the configuration file for future maintenance.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Format README.md to remove extraneous line breaks

Reformatted the README.md file to ensure JSON-like structures within the table are more compact and consistent. The change enhances readability and aligns the code format with the rest of the document.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Simplify Markdown formatting in README

Replaced verbose Markdown syntax with simpler formatting in the README file. This includes changing Markdown for preformatted blocks and lists, ensuring better readability and consistency.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Update module sources to version v8.53.0

This commit updates the 'tls_checker' and 'monitoring_pod_identity' modules in the Terraform configuration to use version v8.53.0 of the modules. This ensures that we are using the latest features and fixes provided in the new module version.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Enable OpsGenie action group only in production environment

Added conditional logic to include OpsGenie action group only if the environment is production. Updated both `01_monitor.tf` and `02_aks.tf` to reflect this change, ensuring non-production environments won't configure OpsGenie.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Remove HTML escape codes from README.md

Cleaned up the README.md by removing unnecessary HTML escape codes. This should make the document easier to read and maintain.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Update precommit version in static analysis workflow

Upgraded the precommit_version from v1.89.1 to v1.96.2 in the static analysis GitHub Action workflow file. This change ensures the latest pre-commit checks are applied to code analysis.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Update README.md with consistent markdown formatting

Replaced HTML tags for line breaks to use self-closing `<br/>` for consistency. This improves readability and ensures uniformity in the documentation formatting.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Update pre-commit Docker image to v1.96.2

This commit updates the pre-commit Docker image to version v1.96.2 with the specified SHA-256 hash. The change ensures compatibility with the latest pre-commit tools and improves the reliability and security of the static analysis workflow.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Update alert metric validation and script enhancements

Added `skip_metric_validation` variable to manage custom metrics. Upgraded Terraform script with enhanced functionalities and updated AKS module to use ref `v8.54.0`.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Add skip_metric_validation to AKS platform metrics

Added the skip_metric_validation variable to multiple metric configurations in 99_locals.tf. This change allows the option to bypass metric validation when needed, providing better flexibility in metric checks.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Disable provider registration for AzureRM

This change sets `skip_provider_registration` to true in the AzureRM provider configuration. It prevents automatic registration of resource providers in Azure, potentially speeding up deployments and reducing unnecessary operations.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

---------

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>