feat: [PAYMCLOUD-253] Add Azure Monitor Workspace and Prometheus integration (#2778)
* Add Azure Monitor Workspace and Prometheus managed addon

Introduced Azure Monitor Workspace for development environments in `next-core`. Enabled Prometheus managed addon in `aks-platform` with integration to the newly-created Monitor Workspace. Updated Terraform module and provider constraints to support these changes.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Upgrade AKS module to version v8.80.0

Updated the Terraform source reference for the AKS module to use version v8.80.0. Adjusted the README accordingly to reflect the version change. This ensures alignment with the latest features and improvements provided in the module.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

---------

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>