chore: [PAYMCLOUD-253] Enable conditional deployment of Prometheus addon for aks weu dev environment (#2779)
* Enable Prometheus addon conditionally for dev environment

Added a conditional `count` to the `prometheus_managed_addon` module, ensuring it is only deployed in the development environment. This change helps avoid unnecessary resource deployment in non-dev environments.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Add conditional logic for dev environment in monitoring setup

Introduce a conditional count for "azurerm_monitor_workspace" to handle dev environment-specific configurations. Adjust references to workspace attributes for compatibility with the count feature.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

---------

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>