feat: [PAYMCLOUD-35] Update AKS module to version v8.84.1 and add new critical node alerts. (#2842)
* Update AKS module to version v8.84.1 with new node critical alerts.

Upgraded the Terraform AKS module from v8.80.0 to v8.84.1 to incorporate the latest improvements and fixes. This ensures better compatibility and reliability for the AKS infrastructure.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Update AKS leonardo module version to v8.84.1 and add new node critical alerts.

Upgraded the Terraform module for the AKS cluster from v8.80.0 to v8.84.1. This ensures compatibility with the latest features and bug fixes provided by the module. No other changes were made to the configuration.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

---------

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>