fix: [PAYMCLOUD-377] Update CI Workflow and Terraform Configurations (#2976)
* Update IaC dependency and adjust TFVars formatting

Upgraded the action reference for modified folder detection to a specific branch for dependency inspection. Additionally, added a comment in the Terraform variables file for clarity.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Update RBAC configuration in 04_rbac.tf

Added a comment line to the RBAC Terraform file, potentially as a placeholder or note for testing purposes. No functional changes to the configuration or resource behavior were made.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Update action to use tagged version in static analysis

Replaced the branch reference with a specific version tag (v1.22.0) for the `get-modified-folders` action. This improves stability and ensures consistent behavior across workflow executions.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* Remove test comment and update folder depth in CI config

The unused test comment in the RBAC Terraform file was removed for clarity. The default folder depth in the static analysis workflow was adjusted from 3 to 2 to refine its scope.

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* wip

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

* wip

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>

---------

Signed-off-by: Fabio Felici <fabio.felici@pagopa.it>